implement rate limiting
default rate limit: 60 reqs / 10 minutes

Use memcache expiry for rate limiting. If cache dumps keys, rate limit can be exceeded, but that's not a huge deal.

Rate limit key is ip + user-agent. A bunch of users behind a router with the same browser will share a key, but this shouldn't be a big issue.

Rate limit only applies to requests that would cost some work:

1. method is cached (most renderers, etc.), and
2. result is not *in* the cache, so it needs to be computed.